fix: move Snapshot to main tests module for shared use
