+++ inline/tests/lexer/snapshot.rs
@@ -1,25 +1,7 @@
-use std::ops::{Deref, DerefMut};
 use unimarkup_commons::test_runner::as_snapshot::AsSnapshot;
 use unimarkup_inline::{Token, TokenKind, Tokens};
 
-/// Wrapper type for implementing the `AsSnapshot` trait.
-/// Integration `tests` is treated as an extra crate, so we can't implement
-/// trait for types where neither are defined in this (`tests`) crate.
-pub(crate) struct Snapshot<T>(pub T);
-
-impl<T> Deref for Snapshot<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<T> DerefMut for Snapshot<T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
+use crate::Snapshot;
 
 impl AsSnapshot for Snapshot<(&str, Tokens)> {
     fn as_snapshot(&self) -> String {
